# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "本地化的 Bangumi 数据服务。利用 Archive 每周导出的 wiki 数据导入本地 SQLite，\n提供条目/人物/角色/章节等 REST 查询接口与单二进制管理接口。\n统一响应包装：成功 {\"ok\": true, \"data\": ...}；失败 {\"ok\": false, \"error\": \"...\"}。\n数据接口均为只读；管理接口（/api/admin/*）需要 X-Admin-Token 请求头（未配置 token 时放行）。\n交互式文档：/docs 或 /swagger/index.html；本文档由 swag 从代码注解自动生成（make docs）。",
        "title": "Bangumi 本地数据搜索 API",
        "contact": {
            "name": "GitHub",
            "url": "https://github.com/nz3u/bangumi-data"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/subjects/search": {
            "get": {
                "description": "q 经符号归一化后匹配原名、中文名与 infobox 别名（「少女歌剧」可命中「少女☆歌剧」）；tag 同时匹配普通标签与元标签，支持多标签组合 `+奇幻,-科幻`（必须包含/必须排除，逗号分隔，无前缀视为必须包含）；sort 留空时有关键词按匹配分级+人气、无关键词按 ID。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "条目"
                ],
                "summary": "条目搜索与筛选",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词（全文搜索，中文子串匹配）",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "作品类型：1 书籍 / 2 动画 / 3 音乐 / 4 游戏 / 6 三次元",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "子类型（平台）ID，取值见 /api/constants",
                        "name": "platform",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签组合，如 +奇幻,-科幻",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "元标签组合（保留兼容，同 tag 语法，仅查元标签）",
                        "name": "meta_tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "排名下界（含）",
                        "name": "rank_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "评分下界（含）",
                        "name": "score_min",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "发售/播出日期下界（YYYY-MM-DD）",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "发售/播出日期上界",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "0=非 R18，1=仅 R18",
                        "name": "nsfw",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "0=非系列，1=系列",
                        "name": "series",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段：id / rank / score / date / favorite，省略=智能排序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "排序方向",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从 1 起",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量，1-200",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.subjectSearchData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/subjects/tags": {
            "get": {
                "description": "按使用次数降序返回候选（前缀命中优先于子串命中）；kind=all 合并普通标签与元标签两张聚合表并按名称去重。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "条目"
                ],
                "summary": "条目标签/元标签实时建议",
                "parameters": [
                    {
                        "enum": [
                            "tag",
                            "meta",
                            "all"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "建议类型",
                        "name": "kind",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "子串过滤（ASCII 大小写不敏感）",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "返回上限，1-5000",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.subjectTagsData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/subjects/{id}": {
            "get": {
                "description": "含基本信息、infobox、简介、双向关联、制作人员、角色与章节数。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "条目"
                ],
                "summary": "条目详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.subjectDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "条目不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.apiEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "api.subjectDetail": {
            "type": "object",
            "properties": {
                "characters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.characterItem"
                    }
                },
                "date": {
                    "type": "string"
                },
                "episode_count": {
                    "type": "integer"
                },
                "favorite": {
                    "$ref": "#/definitions/model.Favorite"
                },
                "id": {
                    "type": "integer"
                },
                "infobox": {
                    "type": "string"
                },
                "meta_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "name_cn": {
                    "type": "string"
                },
                "nsfw": {
                    "type": "boolean"
                },
                "platform": {
                    "type": "integer"
                },
                "platform_name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.relationItem"
                    }
                },
                "score": {
                    "type": "number"
                },
                "score_details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "series": {
                    "type": "boolean"
                },
                "staff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.staffItem"
                    }
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "type": {
                    "type": "integer"
                },
                "type_name": {
                    "type": "string"
                }
            }
        },
...
        "api.subjectSearchData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.subjectBrief"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.subjectTagsData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.tagSuggestItem"
                    }
                }
            }
        },
        "api.tagSuggestItem": {
            "type": "object",
            "properties": {
                "cnt": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.Favorite": {
            "type": "object",
            "properties": {
                "doing": {
                    "type": "integer"
                },
                "done": {
                    "type": "integer"
                },
                "dropped": {
                    "type": "integer"
                },
                "on_hold": {
                    "type": "integer"
                },
                "wish": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        }
    }
}